Move actual request objects to own namespace.
