+++ src/API/Requests/GetMe.php
@@ -1,15 +1,17 @@
 <?php
 
-namespace madpilot78\bottg\API;
+namespace madpilot78\bottg\API\Requests;
 
+use madpilot78\bottg\API\Request;
+use madpilot78\bottg\API\RequestInterface;
 use madpilot78\bottg\Config;
 use madpilot78\bottg\Http\HttpInterface;
 use madpilot78\bottg\Logger;
 
 /**
- * Implements the Telegram Bot API getWebhookInfo.
+ * Implements the Telegram Bot API getMe.
  */
-class GetWebhookInfo extends Request implements RequestInterface
+class GetMe extends Request implements RequestInterface
 {
     /**
      * Constructor, passes correct arguments to upstream constructor.
@@ -29,7 +31,7 @@ class GetWebhookInfo extends Request implements RequestInterface
     ) {
         parent::__construct(
             RequestInterface::GET,
-            'getWebhookInfo',
+            'getMe',
             null,
             $config,
             $logger,
