+++ tests/API/Requests/GetUpdatesTest.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace madpilot78\bottg\tests\API;
+namespace madpilot78\bottg\tests\API\Requests;
 
 use InvalidArgumentException;
-use madpilot78\bottg\API\GetUpdates;
+use madpilot78\bottg\API\Requests\GetUpdates;
 use madpilot78\bottg\API\Response;
 use madpilot78\bottg\Http\HttpInterface;
 use madpilot78\bottg\tests\TestCase;
