+++ tests/API/Requests/GetWebhookInfoTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace madpilot78\bottg\tests\API;
+namespace madpilot78\bottg\tests\API\Requests;
 
-use madpilot78\bottg\API\GetWebhookInfo;
+use madpilot78\bottg\API\Requests\GetWebhookInfo;
 use madpilot78\bottg\API\Response;
 use madpilot78\bottg\Http\HttpInterface;
 use madpilot78\bottg\tests\TestCase;
